dynamic_vino_sample package is renamed to vino_sample for naming consistency
