+++ README.md
@@ -249,7 +249,7 @@ One-step installation scripts are provided for the dependencies' installation. P
 	```
   Run example application with an absolute path of an image on another console:
 	```bash
-	rosrun dynamic_vino_sample image_object_client ~/catkin_ws/src/ros_openvino_toolkit/data/images/car.png
+	rosrun vino_sample image_object_client ~/catkin_ws/src/ros_openvino_toolkit/data/images/car.png
 	```
 * run face detection service sample code input from Image  
   Run image processing service:
@@ -258,7 +258,7 @@ One-step installation scripts are provided for the dependencies' installation. P
 	```
   Run example application with an absolute path of an image on another console:
 	```bash
-	rosrun dynamic_vino_sample image_people_client ~/catkin_ws/src/ros_openvino_toolkit/data/images/team.jpg
+	rosrun vino_sample image_people_client ~/catkin_ws/src/ros_openvino_toolkit/data/images/team.jpg
 	```
 # TODO Features
 * Support **result filtering** for inference process, so that the inference results can be filtered to different subsidiary inference. For example, given an image, firstly we do Object Detection on it, secondly we pass cars to vehicle brand recognition and pass license plate to license number recognition.
