+++ vino_launch/launch/image_reidentification_server_oss.launch
@@ -2,7 +2,7 @@
 
   <arg name="param_file" value="$(find vino_launch)/param/pipeline_reidentification_server_oss.yaml" />
 
-    <node pkg="dynamic_vino_sample" type="image_reidentification_server"
+    <node pkg="vino_sample" type="image_reidentification_server"
         name="image_people_server" output="screen">
     <param name="param_file" value="$(arg param_file)" />
   </node>
