+++ vino_launch/launch/pipeline_segmentation.launch
@@ -11,7 +11,7 @@
          <arg name="color_height" value="$(arg color_height)" />
      </include>
   </group>
-  <node pkg="dynamic_vino_sample" type="pipeline_with_params"
+  <node pkg="vino_sample" type="pipeline_with_params"
         name="pipeline_with_params" output="screen">
     <param name="param_file" value="$(arg param_file)" />
     <remap from ="/openvino_toolkit/segmented_obejcts" to = "/ros_openvino_toolkit/segmented_objects" />
