+++ vino_launch/package.xml
@@ -28,14 +28,14 @@ limitations under the License.
   <build_depend>roslaunch</build_depend>
   <build_depend>dynamic_vino_lib</build_depend>
   <build_depend>vino_param_lib</build_depend>
-  <build_depend>dynamic_vino_sample</build_depend>
+  <build_depend>vino_sample</build_depend>
   <build_depend>usb_cam</build_depend>
   <build_depend>nodelet</build_depend>
 
   <run_depend>roslaunch</run_depend>
   <run_depend>dynamic_vino_lib</run_depend>
   <run_depend>vino_param_lib</run_depend>
-  <run_depend>dynamic_vino_sample</run_depend>
+  <run_depend>vino_sample</run_depend>
   <run_depend>usb_cam</run_depend>
   <run_depend>nodelet</run_depend>
 </package>
