+++ vino_people_msgs/CMakeLists.txt
