+++ vino_people_msgs/msg/Emotion.msg
