+++ vino_people_msgs/msg/EmotionsStamped.msg
