+++ vino_people_msgs/msg/ObjectInMask.msg
