+++ vino_people_msgs/msg/ObjectsInMasks.msg
