+++ vino_people_msgs/msg/ReidentificationStamped.msg
