+++ vino_people_msgs/srv/ObjectsInMasksSrv.srv
