+++ vino_sample/src/image_reidentification_client.cpp
