+++ vino_sample/src/image_segmentation_client.cpp
